% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagRVFL.R
\name{bagRVFL}
\alias{bagRVFL}
\alias{bagRVFL.default}
\title{Bagging random vector functional links.}
\usage{
bagRVFL(X, y, N_hidden, B = 100, lambda = 0, control = list())

\method{bagRVFL}{default}(X, y, N_hidden, B = 100, lambda = 0, control = list())
}
\arguments{
\item{X}{A matrix of observed features used to estimate the parameters of the output layer.}

\item{y}{A vector of observed targets used to estimate the parameters of the output layer.}

\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{B}{The number of bootstrap samples.}

\item{lambda}{The penalisation constant used when training the output layers of each RVFL.}

\item{control}{A list of additional arguments passed to the \link{control_RVFL} function.}
}
\value{
An \link{ERVFL-object}.
}
\description{
Use bootstrap aggregation to reduce the variance of random vector functional link neural network models.
}
\examples{
N <- 2000
p <- 5

X <- matrix(rnorm(N * p), ncol = p)
X[, 1] <- sin(X[, 1])
X[, 2] <- exp(X[, 2])
X[, 3] <- cos(X[, 3])
X[, 4] <- X[, 4] * X[, 4]

beta <- matrix(rnorm(p), ncol = 1) 
y <- X \%*\% beta + rnorm(N, 0, 0.1)

N_hidden <- 100
B <- 100
bagRVFL(X = X, y = y, N_hidden = N_hidden, B = B, lambda = 0.02)
}
