% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R, R/RVFL.R
\name{coef.BRVFL}
\alias{coef.BRVFL}
\alias{coef.RVFL}
\title{Coefficients of the BRVFL object.}
\usage{
\method{coef}{BRVFL}(object, ...)

\method{coef}{RVFL}(object, ...)
}
\arguments{
\item{object}{An RVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Coefficients of the BRVFL object.

Coefficients of the RVFL object.
}
