% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_RVFL.R
\name{coef.RVFL}
\alias{coef.RVFL}
\title{Coefficients of the RVFL-object.}
\usage{
\method{coef}{RVFL}(object, ...)
}
\arguments{
\item{object}{An \link{RVFL-object}.}

\item{...}{Additional arguments.}
}
\value{
The estimated weights of the output-layer.
}
\description{
Coefficients of the RVFL-object.
}
\details{
No additional arguments are used in this instance.
}
