% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_RWNN.R
\name{coef.RWNN}
\alias{coef.RWNN}
\title{Coefficients of the RWNN-object}
\usage{
\method{coef}{RWNN}(object, ...)
}
\arguments{
\item{object}{An \link{RWNN-object}.}

\item{...}{Additional arguments.}
}
\value{
The estimated weights of the output-layer.
}
\description{
Coefficients of the RWNN-object
}
\details{
No additional arguments are used in this instance.
}
