% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_SRVFL.R
\name{coef.SRVFL}
\alias{coef.SRVFL}
\title{Coefficients of an SRVFL object.}
\usage{
\method{coef}{SRVFL}(object, ...)
}
\arguments{
\item{object}{An \link{SRVFL-object}.}

\item{...}{Additional arguments.}
}
\value{
The weights indicated by \code{parameter}.
}
\description{
Coefficients of an SRVFL object.
}
\details{
The only additional argument is \code{parameter} taking the values \code{"W"}, \code{"beta"}, or \code{"sigma"}.
}
