% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_SRWNN.R
\name{coef.SRWNN}
\alias{coef.SRWNN}
\title{Coefficients of an SRWNN-object}
\usage{
\method{coef}{SRWNN}(object, ...)
}
\arguments{
\item{object}{An \link{SRWNN-object}.}

\item{...}{Additional arguments.}
}
\value{
The weights indicated by \code{parameter}.
}
\description{
Coefficients of an SRWNN-object
}
\details{
The only additional argument is \code{parameter} taking the values \code{"W"}, \code{"beta"}, or \code{"sigma"}.
}
