% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_rwnn.R
\name{control_abc_rwnn}
\alias{control_abc_rwnn}
\title{abc_rwnn control function}
\usage{
control_abc_rwnn(
  N_hidden,
  lnorm = NULL,
  bias_hidden = TRUE,
  activation = NULL,
  bias_output = TRUE,
  combine_input = FALSE,
  include_data = TRUE,
  rng = "runif",
  rng_pars = list(min = -5, max = 5),
  N_simulations = 4000,
  metric = NULL,
  epsilon = 0.01,
  trace = NULL
)
}
\arguments{
\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{lnorm}{A string indicating the regularisation norm used when estimating the weights in the output layer (either \code{"l1"} or \code{"l2"}).}

\item{bias_hidden}{A vector of TRUE/FALSE values. The vector should have length 1, or the length should be equal to the number of hidden layers.}

\item{activation}{A vector of strings corresponding to activation functions (see \link{control_rwnn} for possible choices). The vector should have length 1, or the length should be equal to the number of hidden layers.}

\item{bias_output}{TRUE/FALSE: Should a bias be added to the output layer?}

\item{combine_input}{TRUE/FALSE: Should the input and hidden layer be combined for the output of each hidden layer?}

\item{include_data}{TRUE/FALSE: Should the original data be included in the returned object? Note: this should almost always be set to 'TRUE', but can be useful and more memory efficient when bagging or boosting an RWNN.}

\item{rng}{A string indicating the sampling distribution used for generating the weights of the hidden layer (defaults to \code{"runif"}).}

\item{rng_pars}{A list of parameters passed to the \code{rng} function (defaults to \code{list(lower = -1, upper = 1)}).}

\item{N_simulations}{The number of ABC samples drawn from the posterior.}

\item{metric}{A metric function used to compare measured and predicted targets (defaults to the L2-norm). The function needs to take two vectors as input and produce a numeric as output.}

\item{epsilon}{The maximal metric error allowed for the sample to be kept.}

\item{trace}{Show trace every \code{trace} iterations.}
}
\value{
A list of control variables.
}
\description{
A function used to create a control-object for the \link{abc_rwnn} function.
}
