% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_rwnn.R
\name{control_sample_rwnn}
\alias{control_sample_rwnn}
\title{sample_rwnn control function}
\usage{
control_sample_rwnn(
  N_hidden,
  lnorm = NULL,
  bias_hidden = TRUE,
  activation = NULL,
  bias_output = TRUE,
  combine_input = FALSE,
  include_data = TRUE,
  N_simulations = 4000,
  N_burnin = 1000,
  N_resample = 100,
  tau = 0.01,
  method = NULL,
  trace = NULL
)
}
\arguments{
\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{lnorm}{A string indicating the regularisation norm used when estimating the weights in the output layer (either \code{"l1"} or \code{"l2"}).}

\item{bias_hidden}{A vector of TRUE/FALSE values. The vector should have length 1, or the length should be equal to the number of hidden layers.}

\item{activation}{A vector of strings corresponding to activation functions (see \link{control_rwnn} for possible choices). The vector should have length 1, or the length should be equal to the number of hidden layers.}

\item{bias_output}{TRUE/FALSE: Should a bias be added to the output layer?}

\item{combine_input}{TRUE/FALSE: Should the input and hidden layer be combined for the output of each hidden layer?}

\item{include_data}{TRUE/FALSE: Should the original data be included in the returned object? Note: this should almost always be set to 'TRUE', but can be useful and more memory efficient when bagging or boosting an RWNN.}

\item{N_simulations}{The number of Metropolis-Hastings samples drawn from the posterior.}

\item{N_burnin}{The number of samples used as burn-in during Metropolis-Hastings sampling.}

\item{N_resample}{The number of samples drawn during re-sampling after Metropolis-Hastings.}

\item{tau}{The step-size used when generating new weights from the previous step in Metropolis-Hastings sampler.}

\item{method}{String indicating the method used after Metropolis-Hastings sampling (see details).}

\item{trace}{Show trace every \code{trace} iterations.}
}
\value{
A list of control variables.
}
\description{
A function used to create a control-object for the \link{sample_rwnn} function.
}
\details{
The current choices of \code{method} are:
\describe{
    \item{\code{"map"}}{Only the maximum a posterior (MAP) estimate of the weights is returned as an \link{RWNN-object}.}
    \item{\code{"stack"}}{The weights are re-sampled (with replacement) weighted using the unnormalised posterior, creating a stacking based \link{ERWNN-object}.}
    \item{\code{"posterior"}}{The posterior distributions of the weights are returned.}
}
}
