% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRVFL.R
\name{control_sampleRVFL}
\alias{control_sampleRVFL}
\title{sampleRVFL control function.}
\usage{
control_sampleRVFL(
  N_simulations = 4000,
  N_burnin = 1000,
  N_resample = 100,
  tau = 0.01,
  method = NULL,
  trace = NULL
)
}
\arguments{
\item{N_simulations}{The number of Metropolis-Hastings samples drawn from the posterior.}

\item{N_burnin}{The number of samples used as burn-in during Metropolis-Hastings sampling.}

\item{N_resample}{The number of samples drawn during re-sampling after Metropolis-Hastings.}

\item{tau}{The step-size used when generating new weights from the previous step in Metropolis-Hastings sampler.}

\item{method}{String indicating the method used after Metropolis-Hastings sampling (see details).}

\item{trace}{Show trace every \code{trace} iterations.}
}
\value{
A list of control variables.
}
\description{
A function used to create a control-object for the \link{sampleRVFL} function.
}
\details{
The current choices of \code{method} are:
\describe{
    \item{\code{"map"}}{Only the maximum a posterior (MAP) estimate of the weights is returned.}
    \item{\code{"resampling"}}{The weights are resampled (with replacement) weighted using the unnormalised posterior.}
}
}
