% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVFL.R
\name{Errors}
\alias{Errors}
\alias{mse}
\alias{rmse}
\alias{mae}
\alias{mape}
\title{Error functions}
\usage{
mse(object, X, y)

rmse(object, X, y)

mae(object, X, y)

mape(object, X, y)
}
\arguments{
\item{object}{A model object}

\item{X}{A matrix of observed features.}

\item{y}{A vector of observed targets.}
}
\value{
Error for the provided data-set.
}
\description{
Simple error functions for finding MSE, RMSE, MAE, and MAPE.
}
