% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVFL-package.R
\name{ERVFL-object}
\alias{ERVFL-object}
\title{An ERVFL-object}
\description{
An ERVFL-object is a list containing the following:
\describe{
    \item{\code{data}}{The original data used to estimate the weights.}
    \item{\code{RVFLmodels}}{A list with each element being an \link{RVFL-object}.}
    \item{\code{weights}}{A vector of ensemble weights.}
    \item{\code{method}}{A string indicating the method.}
}
}
