% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_ERWNN.R
\name{estimate_weights}
\alias{estimate_weights}
\alias{estimate_weights.ERWNN}
\title{Estimate ensemble weights for an ERWNN-object}
\usage{
estimate_weights(object, X_val = NULL, y_val = NULL)

\method{estimate_weights}{ERWNN}(object, X_val = NULL, y_val = NULL)
}
\arguments{
\item{object}{An \link{ERWNN-object}.}

\item{X_val}{The validation feature set.}

\item{y_val}{The validation target set.}
}
\value{
An \link{ERWNN-object}.
}
\description{
Estimate ensemble weights for an \link{ERWNN-object}.
}
\examples{
##
N <- dim(example_data)[1]
train_fraction <- 0.7
train_index <- sample(N, train_fraction * N)

data_train <- example_data[train_index, ]
data_val <- example_data[-train_index, ]

###
N_hidden <- 10
B <- 100
mm <- bag_rwnn(y ~ ., data = data_train, 
               N_hidden = N_hidden, lambda = 0.2, B = B)

###
estimate_weights(mm, X_val = data_val[, -1], y_val = data_val[, 1])
}
