% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWNN-package.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data}
\format{
An object of class \code{data.frame} with 2000 rows and 6 columns.
}
\usage{
example_data
}
\description{
Data generated using a simple non-linear function with 5 inputs and 1 output.
}
\keyword{datasets}
