% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krvfl.R
\name{krvfl}
\alias{krvfl}
\title{Simple and efficient implementation using the kernel trick on concatenated features}
\usage{
krvfl(
  x,
  y,
  lambda = 0.1,
  nb_hidden = 100L,
  activation = c("relu", "sigmoid", "tanh", "linear"),
  sigma = 1,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{Matrix of predictors (n x p)}

\item{y}{Response vector (n x 1) or matrix (n x m)}

\item{lambda}{Regularization parameter}

\item{nb_hidden}{Number of hidden units}

\item{activation}{Activation function}

\item{sigma}{Scale parameter for weights}

\item{seed}{Random seed}

\item{...}{Additional arguments}
}
\value{
krvfl object
}
\description{
Simple and efficient implementation using the kernel trick on concatenated features
}
