% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R
\name{plot.BRVFL}
\alias{plot.BRVFL}
\title{Diagnostic-plots of an BRVFL-object.}
\usage{
\method{plot}{BRVFL}(x, ...)
}
\arguments{
\item{x}{A BRVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Diagnostic-plots of an BRVFL-object.
}
\details{
The additional arguments used by the function are '\code{X_val}' and '\code{y_val}', i.e. the features and targets of the validation-set. These are helpful when analysing whether overfitting of model has occurred.
}
