% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_ERVFL.R
\name{plot.ERVFL}
\alias{plot.ERVFL}
\title{Diagnostic-plots of an ERVFL-object.}
\usage{
\method{plot}{ERVFL}(x, ...)
}
\arguments{
\item{x}{An \link{ERVFL-object}.}

\item{...}{Additional arguments.}
}
\description{
Diagnostic-plots of an ERVFL-object.
}
\details{
The additional arguments used by the function are '\code{X_val}' and '\code{y_val}', i.e. the features and targets of the validation-set. These are helpful when analysing whether overfitting of model has occurred.
}
