% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_ERWNN.R
\name{plot.ERWNN}
\alias{plot.ERWNN}
\title{Diagnostic-plots of an ERWNN-object}
\usage{
\method{plot}{ERWNN}(x, ...)
}
\arguments{
\item{x}{An \link{ERWNN-object}.}

\item{...}{Additional arguments.}
}
\description{
Diagnostic-plots of an ERWNN-object
}
\details{
The additional arguments used by the function are '\code{X_val}' and '\code{y_val}', i.e. the features and targets of the validation-set. These are helpful when analysing whether overfitting of model has occurred.
}
