% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVFL.R
\name{plot.RVFL}
\alias{plot.RVFL}
\title{Diagnostic-plots of an RVFL-object.}
\usage{
\method{plot}{RVFL}(object, ...)
}
\arguments{
\item{object}{An RVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Diagnostic-plots of an RVFL-object.
}
\details{
The additional arguments used by the function are '\code{testing_X}' and '\code{testing_y}', i.e. the features and targets of the testing-set. These are helpful when analysing whether overfitting of model has occured.
}
