% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_RWNN.R
\name{plot.RWNN}
\alias{plot.RWNN}
\title{Diagnostic-plots of an RWNN-object}
\usage{
\method{plot}{RWNN}(x, ...)
}
\arguments{
\item{x}{An \link{RWNN-object}.}

\item{...}{Additional arguments.}
}
\description{
Diagnostic-plots of an RWNN-object
}
\details{
The only additional argument used by the function is '\code{newdata}', which expects a matrix with the same number of features (columns) as in the original data.
}
