% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_SRWNN.R
\name{plot.SRWNN}
\alias{plot.SRWNN}
\title{Diagnostic-plots of an SRWNN-object}
\usage{
\method{plot}{SRWNN}(x, ...)
}
\arguments{
\item{x}{An \link{SRWNN-object}.}

\item{...}{Additional arguments.}
}
\description{
Diagnostic-plots of an SRWNN-object
}
\details{
The additional arguments used by the plot.SRWNN function are: 
\describe{
   \item{\code{parameter}}{A character indicating the parameter, either "W", "beta", or "sigma".}
   \item{\code{N_samples}}{The number of samples drawn from the posterior.}
   \item{\code{index}}{A vector indicating the layer, neuron, and connection to show when \code{parameter = "W"}.}
   \item{\code{breaks}}{The number of breaks used in the histogram when \code{paramter = "W"} or \code{"sigma"} (default: \code{"fd"}).}
}
}
