% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R
\name{predict.BRVFL}
\alias{predict.BRVFL}
\title{Predicting targets of an BRVFL object.}
\usage{
\method{predict}{BRVFL}(object, ...)
}
\arguments{
\item{object}{A BRVFL-object.}

\item{...}{Additional arguments.}
}
\value{
Depended on \code{type}: 
\describe{
    \item{\code{"all"}}{A matrix where every column contains the predicted values corresponding to each of the boostrapped models.}
    \item{\code{"sd"}}{A vector containing the standard deviation of each prediction taken across the bootstrap samples.}
    \item{\code{"mean"}}{A vector containing the weighted (using the \code{weights} element of the \link{BRVFL}-object) sum each observation taken across the bootstrap samples.}
}
}
\description{
Predicting targets of an BRVFL object.
}
\details{
The additional argument '\code{newdata}' and '\code{type}' can be specified:
\describe{
  \item{\code{newdata}}{Expects a matrix the same number of features (columns) as in the original data.}
  \item{\code{type}}{Takes values \code{"all"}, \code{"sd"}, and \code{"mean"} (default), returning a full matrix of predictions for all bootstrap samples, the standard deviation of each predicted observation across bootstrap samples, and the average value of each prediction across the bootstrap samples, respectively.}
}
}
