% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krvfl.R
\name{predict.krvfl}
\alias{predict.krvfl}
\title{Predict method for krvfl objects}
\usage{
\method{predict}{krvfl}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A krvfl object from krvfl()}

\item{newdata}{Matrix of new predictors (m x p)}

\item{...}{Additional arguments (unused)}
}
\value{
Predicted values (m x 1) or (m x q) matrix
}
\description{
Predict method for krvfl objects
}
