% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVFL.R
\name{predict.RVFL}
\alias{predict.RVFL}
\title{Predicting targets of an RVFL object.}
\usage{
\method{predict}{RVFL}(object, ...)
}
\arguments{
\item{object}{An RVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Predicting targets of an RVFL object.
}
