% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_RVFL.R
\name{predict.RVFL}
\alias{predict.RVFL}
\title{Predicting targets of an RVFL-object.}
\usage{
\method{predict}{RVFL}(object, ...)
}
\arguments{
\item{object}{An \link{RVFL-object}.}

\item{...}{Additional arguments.}
}
\value{
A vector of predicted targets.
}
\description{
Predicting targets of an RVFL-object.
}
\details{
The only additional argument used by the function is '\code{newdata}', which expects a matrix with the same number of features (columns) as in the original data.
}
