% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_RWNN.R
\name{predict.RWNN}
\alias{predict.RWNN}
\title{Predicting targets of an RWNN-object}
\usage{
\method{predict}{RWNN}(object, ...)
}
\arguments{
\item{object}{An \link{RWNN-object}.}

\item{...}{Additional arguments.}
}
\value{
A vector of predicted targets.
}
\description{
Predicting targets of an RWNN-object
}
\details{
The only additional argument used by the function is '\code{newdata}', which expects a matrix with the same number of features (columns) as in the original data.
}
