% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_SRWNN.R
\name{predict.SRWNN}
\alias{predict.SRWNN}
\title{Predicting targets of an SRWNN-object}
\usage{
\method{predict}{SRWNN}(object, ...)
}
\arguments{
\item{object}{An \link{SRWNN-object}.}

\item{...}{Additional arguments.}
}
\value{
Either a vector or matrix dependent '\code{type}' (see details).
}
\description{
Predicting targets of an SRWNN-object
}
\details{
The additional argument '\code{newdata}' and '\code{type}' can be specified, as follows:
\describe{
  \item{\code{newdata}}{Expects a matrix the same number of features (columns) as in the original data.}
  \item{\code{N_samples}}{The number of samples drawn from the posterior.}
  \item{\code{type}}{Takes values \code{"all"}, \code{"mean"} (default), and \code{"ci"}, returning a matrix of predictions for all posterior samples, a vector of mean prediction across samples, and a matrix of prediction intervals across the samples.}
  \item{\code{p}}{A numeric indicating the credibility level when \code{type = "ci"}.}
}
}
