% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_network.R
\name{reduce_network}
\alias{reduce_network}
\alias{reduce_network.RWNN}
\alias{reduce_network.ERWNN}
\title{Reduce random weight neural network}
\usage{
reduce_network(object, type = "FI")

\method{reduce_network}{RWNN}(object, type = "FI")

\method{reduce_network}{ERWNN}(object, type = "FI")
}
\arguments{
\item{object}{An \link{RWNN-object} or \link{ERWNN-object}.}

\item{type}{A string setting the method used to reduce the network.}
}
\value{
A reduce \link{RWNN-object}.
}
\description{
.
}
