% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_network.R
\name{reduce_retrain_network}
\alias{reduce_retrain_network}
\alias{reduce_retrain_network.RWNN}
\alias{reduce_retrain_network.ERWNN}
\title{Reduce the weights of a random weight neural network.}
\usage{
reduce_retrain_network(object, ...)

\method{reduce_retrain_network}{RWNN}(object, ...)

\method{reduce_retrain_network}{ERWNN}(object, ...)
}
\arguments{
\item{object}{An \link{RWNN} or \link{ERWNN}-object.}

\item{...}{Arguments passed to the \link{reduce_network}-function.}
}
\value{
A reduced \link{RWNN} or \link{ERWNN}-object.
}
\description{
Methods for weight and neuron pruning in random weight neural networks.
}
