% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_ERWNN.R
\name{residuals.ERWNN}
\alias{residuals.ERWNN}
\title{Residuals of the ERWNN-object}
\usage{
\method{residuals}{ERWNN}(object, ...)
}
\arguments{
\item{object}{An \link{ERWNN-object}.}

\item{...}{Additional arguments.}
}
\value{
A vector of raw residuals between the predicted (using \code{type = "mean"}) and observed targets.
}
\description{
Residuals of the ERWNN-object
}
\details{
No additional arguments are used in this instance.
}
