% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R, R/RVFL.R
\name{residuals.BRVFL}
\alias{residuals.BRVFL}
\alias{residuals.RVFL}
\title{Residuals of the BRVFL object.}
\usage{
\method{residuals}{BRVFL}(object, ...)

\method{residuals}{RVFL}(object, ...)
}
\arguments{
\item{object}{An RVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Residuals of the BRVFL object.

Residuals of the RVFL object.
}
