% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_RWNN.R
\name{residuals.RWNN}
\alias{residuals.RWNN}
\title{Residuals of the RWNN-object}
\usage{
\method{residuals}{RWNN}(object, ...)
}
\arguments{
\item{object}{An \link{RWNN-object}.}

\item{...}{Additional arguments.}
}
\value{
A vector of residuals of the desired '\code{type}' (see details).
}
\description{
Residuals of the RWNN-object
}
\details{
Besides the arguments passed to the '\code{predict}' function, the argument '\code{type}' can be supplied defining the type of residual returned by the function. Currently only \code{"rs"} (standardised residuals), and \code{"raw"} (default) are implemented.
}
