% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFRVFL.R
\docType{package}
\name{RFRVFL}
\alias{RFRVFL}
\title{RFRVFL: Random Forest Random Vector Functional Link Neural Network Models}
\description{
This package allows for simple creation, estimation, and prediction of random vector functional link neural networks. Furthermore, it incorporates a random forest style boosting into the RVFL framework, yielding more robust estimation and prediction.
}
\author{
Søren B. Vilsen <svilsen@math.aau.dk>
}
