% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVFL-package.R
\name{RVFL-object}
\alias{RVFL-object}
\title{An RVFL-object}
\description{
An RVFL-object is a list containing the following:
\describe{
    \item{\code{data}}{The original data used to estimate the weights.}
    \item{\code{N_hidden}}{The vector of neurons in each layer.}
    \item{\code{activation}}{The vector of the activation functions used in each layer.}
    \item{\code{lambda}}{The penalisation constant used when estimating the output weights.}
    \item{\code{Bias}}{The \code{TRUE/FALSE} bias vectors set by the control function for both hidden layers, and the output layer.}
    \item{\code{Weights}}{The weigths of the neural network, split into random (stored in hidden) and estimated (stored in output) weights.}
    \item{\code{Sigma}}{The standard deviation of the corresponding linear model.}
    \item{\code{Combined}}{A \code{TRUE/FALSE} stating whether the direct links were made to the input.}
}
}
