% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVFL-package.R
\docType{package}
\name{RVFL-package}
\alias{RVFL-package}
\title{RVFL: Random Vector Functional Link Neural Network Models}
\description{
Creation, estimation, and prediction of random vector functional link neural networks (RVFL), including popular variants like extreme learning machines (ELM), sparse RVFL (spRVFL), and deep RVFL (dRVFL). It further allows for the creation of ensemble RVFLs like bagging RVFL (bagRVFL), boosting RVFL (boostRVFL), stacking RVFL (stackRVFL), and ensemble deep RVFL (edRVFL).
}
\author{
Søren B. Vilsen <svilsen@math.aau.dk>
}
