% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFRVFL.R
\name{set_weights}
\alias{set_weights}
\alias{set_weights.RFRVFL}
\title{Set ensemble weights for an RFRVFL-object.}
\usage{
set_weights(object, weights = NULL)

\method{set_weights}{RFRVFL}(object, weights = NULL)
}
\arguments{
\item{object}{An RFRVFL-object.}

\item{weights}{A vector of ensemble weights.}
}
\value{
An RFRVFL-object.
}
\description{
Manually set ensemble weights for an RFRVFL-object.
}
