% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krvfl.R
\name{summary.krvfl}
\alias{summary.krvfl}
\title{Summary method for krvfl objects}
\usage{
\method{summary}{krvfl}(object, ...)
}
\arguments{
\item{object}{A krvfl object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for krvfl objects
}
