% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvfl2.R
\name{summary.rvfl2}
\alias{summary.rvfl2}
\title{Summary of RVFL2 Model}
\usage{
\method{summary}{rvfl2}(object)
}
\arguments{
\item{object}{An object of class \code{rvfl2} (the fitted model).}
}
\value{
None (prints summary information to the console).
}
\description{
Displays a summary of the RVFL2 model, including the key hyperparameters.
}
