% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_hyperparameters.R
\name{tune_hyperparameters}
\alias{tune_hyperparameters}
\alias{tune_hyperparameters.formula}
\title{Hyper-parameter tuning}
\usage{
tune_hyperparameters(
  formula,
  method,
  data = NULL,
  folds = 10,
  hyperparameters = list(),
  control = list(),
  trace = 0
)

\method{tune_hyperparameters}{formula}(
  formula,
  method,
  data = NULL,
  folds = 10,
  hyperparameters = list(),
  control = list(),
  trace = 0
)
}
\arguments{
\item{formula}{A \link{formula} specifying features and targets used to estimate the parameters of the output layer.}

\item{method}{The RWNN method in need to hyper parameter optimisation.}

\item{data}{A data-set (either a \link{data.frame} or a \link[tibble]{tibble}) used to estimate the parameters of the output layer.}

\item{folds}{The number of folds used in k-fold cross-validation.}

\item{hyperparameters}{A list of sequences of hyper-parameters.}

\item{control}{A list of additional arguments passed to the \link{control_rwnn} function.}

\item{trace}{A numeric indicating how often a trace of should be shown (default is 0).}
}
\value{
Either an \link{RWNN-object} or \link{ERWNN-object}.
}
\description{
Simple function for hyper-parameter tuning using k-fold cross-validation.
}
\examples{
hyperparameters <- list(
    N_hidden = list(c(100, 100, 50, 50), c(100, 10, 10), c(100)), 
    lambda = exp(seq(-8, 2))
)

folds <- 20

\dontrun{
tune_hyperparameters(
    y ~ ., data = example_data, 
    rwnn, 
    folds = folds, hyperparameters = hyperparameters, 
    trace = 1
)
}
}
