% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_rwnn.R
\name{abc_rwnn}
\alias{abc_rwnn}
\title{Approximate Bayesian computation random weight neural networks}
\usage{
abc_rwnn(X, y, N_hidden = c(), lambda = NULL, control = list())
}
\arguments{
\item{X}{A matrix of observed features used to estimate the parameters of the output layer.}

\item{y}{A vector of observed targets used to estimate the parameters of the output layer.}

\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{lambda}{The penalisation constant used when training the output layers of the RWNN.}

\item{control}{A list of additional arguments passed to the \link{control_abc_rwnn} function (includes arguments passed to the \link{control_rwnn} function.).}
}
\description{
Uses approximate Bayesian computation to sample the distribution of the hidden layers in random weight neural network models.
}
