% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_rwnn.R
\name{bag_rwnn}
\alias{bag_rwnn}
\alias{bag_rwnn.formula}
\title{Bagging random weight neural networks}
\usage{
bag_rwnn(
  formula,
  data = NULL,
  n_hidden = c(),
  lambda = NULL,
  B = 100,
  method = NULL,
  type = NULL,
  control = list()
)

\method{bag_rwnn}{formula}(
  formula,
  data = NULL,
  n_hidden = c(),
  lambda = NULL,
  B = 100,
  method = NULL,
  type = NULL,
  control = list()
)
}
\arguments{
\item{formula}{A \link{formula} specifying features and targets used to estimate the parameters of the output layer.}

\item{data}{A data-set (either a \link{data.frame} or a \link[tibble]{tibble}) used to estimate the parameters of the output layer.}

\item{n_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{lambda}{The penalisation constant used when training the output layers of each RWNN.}

\item{B}{The number of bootstrap samples.}

\item{method}{The penalisation type passed to \link{ae_rwnn}. Set to \code{NULL} (default), \code{"l1"}, or \code{"l2"}. If \code{NULL}, the \link{rwnn} is used as the base learner.}

\item{type}{A string indicating whether this is a regression or classification problem.}

\item{control}{A list of additional arguments passed to the \link{control_rwnn} function.}
}
\value{
An \link{ERWNN-object}.
}
\description{
Use bootstrap aggregation to reduce the variance of random weight neural network models.
}
\examples{
n_hidden <- 50
B <- 100
lambda <- 0.01

# Regression
\dontrun{
m <- bag_rwnn(y ~ ., data = example_data, n_hidden = n_hidden, lambda = lambda, B = B)
}

# Classification
\dontrun{
m <- bag_rwnn(I(y > 15) ~ ., data = example_data, n_hidden = n_hidden, lambda = lambda, B = B)
}

}
