% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R
\name{coef.BRVFL}
\alias{coef.BRVFL}
\title{Coefficients of the BRVFL object.}
\usage{
\method{coef}{BRVFL}(object, ...)
}
\arguments{
\item{object}{A BRVFL-object.}

\item{...}{Additional arguments.}
}
\value{
Depended on \code{type}: 
\describe{
    \item{\code{"all"}}{A matrix where every column contains the parameters of the output-layer of corresponding boostrap sample.}
    \item{\code{"sd"}}{A vector containing the standard deviation of each parameter taken across the bootstrap samples.}
    \item{\code{"mean"}}{A vector containing the average value of each parameter taken across the bootstrap samples.}
}
}
\description{
Coefficients of the BRVFL object.
}
\details{
The additional argument '\code{type}' can be supplied with values \code{"all"}, \code{"sd"}, and \code{"mean"} (default), returning the full list of coefficients for all bootstrap samples, the standard deviation of each coefficient across bootstrap samples, and the average value of each coefficient across bootstrap samples, respectively.
}
