% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFRVFL.R, R/RVFL.R
\name{coef.RFRVFL}
\alias{coef.RFRVFL}
\alias{coef.RVFL}
\title{Coefficients of the RFRVFL object.}
\usage{
\method{coef}{RFRVFL}(object, ...)

\method{coef}{RVFL}(object, ...)
}
\arguments{
\item{object}{An RVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Coefficients of the RFRVFL object.

Coefficients of the RVFL object.
}
