% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVFL.R
\name{control_RVFL}
\alias{control_RVFL}
\title{RVFL control function}
\usage{
control_RVFL(
  hidden_bias = TRUE,
  activation = NULL,
  weight_method = "normal",
  weight_mean = 0,
  weight_sd = 1,
  output_bias = TRUE,
  combine_input = FALSE,
  trace = 0
)
}
\arguments{
\item{hidden_bias}{A vector of TRUE/FALSE values. The vector should have length 1, or the length should be equal to the number of hidden layers.}

\item{activation}{A vector of activation functions.}

\item{weight_method}{The method used for generating random weights.}

\item{weight_mean}{The mean of the randomly initialised weights.}

\item{weight_sd}{The standard deviation of the randomly initialised weights.}

\item{trace}{An integer passed to the optimiser. If set larger than 0, a trace of the optimisation iterations is shown every '\code{trace}' iterations.}
}
\value{
A list of control variables.
}
\description{
A function used to create a control-object for the \link{RVFL} function.
}
