% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVFL.R
\name{control_RVFL}
\alias{control_RVFL}
\title{RVFL control function}
\usage{
control_RVFL(
  bias_hidden = TRUE,
  activation = NULL,
  weight_method = "normal",
  weight_mean = 0,
  weight_sd = 1,
  bias_output = TRUE,
  combine_input = FALSE
)
}
\arguments{
\item{bias_hidden}{A vector of TRUE/FALSE values. The vector should have length 1, or the length should be equal to the number of hidden layers.}

\item{activation}{A vector of activation functions (NOT IMPLEMENTED -- IN THIS VERSION ALL ACTIVATIONS ARE SIGMOID).}

\item{weight_method}{The method used for generating random weights.}

\item{weight_mean}{The mean of the randomly initialised weights.}

\item{weight_sd}{The standard deviation of the randomly initialised weights.}

\item{bias_output}{TRUE/FALSE: Should a bias be added to the output layer?}

\item{combine_input}{TRUE/FALSE: Should the input and hidden layer be combined for the output layer?}
}
\value{
A list of control variables.
}
\description{
A function used to create a control-object for the \link{RVFL} function.
}
