% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_rwnn.R
\name{ed_rwnn}
\alias{ed_rwnn}
\alias{ed_rwnn.formula}
\title{Ensemble deep random weight neural networks}
\usage{
ed_rwnn(formula, data = NULL, N_hidden, lambda = 0, control = list())

\method{ed_rwnn}{formula}(formula, data = NULL, N_hidden, lambda = 0, control = list())
}
\arguments{
\item{formula}{A \link{formula} specifying features and targets used to estimate the parameters of the output layer.}

\item{data}{A data-set (either a \link{data.frame} or a \link[tibble]{tibble}) used to estimate the parameters of the output layer.}

\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{lambda}{The penalisation constant used when training the output layers of each RWNN.}

\item{control}{A list of additional arguments passed to the \link{control_rwnn} function.}
}
\value{
An \link{ERWNN-object}.
}
\description{
Use multiple layers to create deep ensemble random weight neural network models.
}
\examples{
N_hidden <- rep(10, 15)
ed_rwnn(y ~ ., data = example_data, 
        N_hidden = N_hidden, lambda = 0.2)
}
