% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edRVFL.R
\name{edRVFL}
\alias{edRVFL}
\alias{edRVFL.default}
\title{Ensemble deep random vector functional links}
\usage{
edRVFL(X, y, N_hidden, lambda = 0, control = list())

\method{edRVFL}{default}(X, y, N_hidden, lambda = 0, control = list())
}
\arguments{
\item{X}{A matrix of observed features used to estimate the parameters of the output layer.}

\item{y}{A vector of observed targets used to estimate the parameters of the output layer.}

\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{lambda}{The penalisation constant used when training the output layers of each RVFL.}

\item{control}{A list of additional arguments passed to the \link{control_RVFL} function.}
}
\value{
An \link{ERVFL-object}.
}
\description{
Use multiple layers to create deep ensemble random vector functional link neural network models.
}
\examples{
N <- 2000
p <- 5

s <- seq(0, pi, length.out = N)
X <- matrix(NA, ncol = p, nrow = N)
X[, 1] <- sin(s)
X[, 2] <- cos(s)
X[, 3] <- s
X[, 4] <- s^2
X[, 5] <- s^3

beta <- matrix(rnorm(p), ncol = 1) 
y <- X \%*\% beta + rnorm(N, 0, 1)

N_hidden <- rep(10, 15)
edRVFL(X = X, y = y, N_hidden = N_hidden, lambda = 0.2)
}
