% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWNN.R
\name{elm}
\alias{elm}
\alias{elm.default}
\alias{elm.formula}
\title{Extreme learning machine}
\usage{
elm(X, y, formula, data, N_hidden = c(), lambda = NULL, control = list())

\method{elm}{default}(X, y, N_hidden, lambda = 0, control = list())

\method{elm}{formula}(formula, data, N_hidden, lambda = 0, control = list())
}
\arguments{
\item{X}{A matrix of observed features used to train the parameters of the output layer.}

\item{y}{A vector of observed targets used to train the parameters of the output layer.}

\item{formula}{A \link{formula} specifying features and targets used to estimate the parameters of the output layer.}

\item{data}{A data-set (either a \link{data.frame} or a \link{tibble}) used to estimate the parameters of the output layer.}

\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{lambda}{The penalisation constant used when training the output layer.}

\item{control}{A list of additional arguments passed to the \link{control_rwnn} function.}
}
\value{
An \link{RWNN-object}.
}
\description{
Set-up and estimate weights of an extreme learning machine.
}
\details{
The function \code{elm} is a wrapper for the general \code{RWNN} function without the link between features and targets.
}
