% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{estimate_weights}
\alias{estimate_weights}
\alias{estimate_weights.ERWNN}
\title{Estimate ensemble weights of an ERWNN-object}
\usage{
estimate_weights(object, data_val = NULL)

\method{estimate_weights}{ERWNN}(object, data_val = NULL)
}
\arguments{
\item{object}{An \link{ERWNN-object}.}

\item{data_val}{A data.frame or tibble containing the validation-set.}
}
\value{
An \link{ERWNN-object}.
}
\description{
Estimate ensemble weights of an \link{ERWNN-object}.
}
\examples{
N <- nrow(example_data)

train_fraction <- 0.7
train_index <- sample(N, train_fraction * N)

data_train <- example_data[train_index, ]
data_val <- example_data[-train_index, ]

###
n_hidden <- 10
B <- 100
lambda <- 0.1

mm <- boost_rwnn(y ~ ., data = data_train, n_hidden = n_hidden, lambda = lambda, B = B)

###
estimate_weights(mm, data_val = data_val)
}
