% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFRVFL.R
\name{estimate_weights}
\alias{estimate_weights}
\alias{estimate_weights.RFRVFL}
\title{Estimate ensemble weights for an RFRVFL-object.}
\usage{
estimate_weights(object, validation_X = NULL, validation_y = NULL)

\method{estimate_weights}{RFRVFL}(object, validation_X = NULL, validation_y = NULL)
}
\arguments{
\item{object}{An RFRVFL-object.}

\item{validation_X}{The validation feature set.}

\item{validation_y}{The validation target set.}
}
\value{
An RFRVFL-object.
}
\description{
Estimate ensemble weights for an RFRVFL-object.
}
