% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R
\name{estimate_weights}
\alias{estimate_weights}
\alias{estimate_weights.BRVFL}
\title{Estimate ensemble weights for an BRVFL-object.}
\usage{
estimate_weights(object, validation_X = NULL, validation_y = NULL)

\method{estimate_weights}{BRVFL}(object, validation_X = NULL, validation_y = NULL, trace = 0)
}
\arguments{
\item{object}{An BRVFL-object.}

\item{validation_X}{The validation feature set.}

\item{validation_y}{The validation target set.}

\item{trace}{Numeric passed to \link{solnp}. Default is set to 0.}
}
\value{
An BRVFL-object.
}
\description{
Estimate ensemble weights for an BRVFL-object.
}
