% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R
\name{estimate_weights}
\alias{estimate_weights}
\alias{estimate_weights.BRVFL}
\title{Estimate ensemble weights for an BRVFL-object.}
\usage{
estimate_weights(object, validation_X = NULL, validation_y = NULL)

\method{estimate_weights}{BRVFL}(object, validation_X = NULL, validation_y = NULL, trace = 0)
}
\arguments{
\item{object}{An BRVFL-object.}

\item{validation_X}{A matrix of observed features used to estimate the weights.}

\item{validation_y}{A vector of observed targets used to estimate the weights.}

\item{trace}{The trace of \link{solnp} are printed every 'trace' number of iteration (default 0).}
}
\value{
An BRVFL-object.
}
\description{
Estimate ensemble weights for an BRVFL-object.
}
\examples{
### 
N <- 200
p <- 5

X <- matrix(rnorm(N * p), ncol = p) 
beta <- matrix(runif(p), ncol = 1) 
y <- X \%*\% beta + rnorm(p)

proportion_training <- 0.7
proportion_validation <- 1L - proportion_training
index_training <- sample(N, 0.7 * N)

training_X <- X[index_training, ]
validation_X <- X[-index_training, ]

training_y <- matrix(y[index_training, ], ncol = 1)
validation_y <- matrix(y[-index_training, ], ncol = 1)

###
N_hidden <- c(10, 2, 4, 2)
B <- 100
mm <- BRVFL(X = training_X, y = training_y, N_hidden = N_hidden, 
            B = B, combine_input = FALSE)

###
estimate_weights(mm, validation_X = validation_X, validation_y = validation_y)
}
