% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUX_ERVFL.R
\name{estimate_weights}
\alias{estimate_weights}
\alias{estimate_weights.ERVFL}
\title{Estimate ensemble weights for an ERVFL-object.}
\usage{
estimate_weights(object, X_val = NULL, y_val = NULL, trace = 0)

\method{estimate_weights}{ERVFL}(object, X_val = NULL, y_val = NULL, trace = 0)
}
\arguments{
\item{object}{An ERVFL-object.}

\item{X_val}{The validation feature set.}

\item{y_val}{The validation target set.}

\item{trace}{The trace of \link{solnp} are printed every '\code{trace}' number of iteration (default 0).}
}
\value{
An ERVFL-object.
}
\description{
Estimate ensemble weights for an ERVFL-object.
}
\examples{
### 
N <- 200
p <- 5

X <- matrix(rnorm(N * p), ncol = p) 
beta <- matrix(runif(p), ncol = 1) 
y <- X \%*\% beta + rnorm(p)

##
train_fraction <- 0.7
train_index <- sample(N, train_fraction * N)

X_train <- X[train_index, ]
X_val <- X[-train_index, ]

y_train <- matrix(y[train_index, ], ncol = 1)
y_val <- matrix(y[-train_index, ], ncol = 1)

###
N_hidden <- 10
B <- 100
mm <- bagRVFL(X = X_train, y = y_train, N_hidden = N_hidden,
              B = B, lambda = 0.2, N_features = NULL)

###
estimate_weights(mm, X_val = X_val, y_val = y_val)
}
