% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUX_RVFL.R
\name{plot.RVFL}
\alias{plot.RVFL}
\title{Diagnostic-plots of an RVFL-object.}
\usage{
\method{plot}{RVFL}(x, ...)
}
\arguments{
\item{x}{An RVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Diagnostic-plots of an RVFL-object.
}
\details{
The additional arguments used by the function are '\code{X_val}' and '\code{y_val}', i.e. the features and targets of the validation-set. These are helpful when analysing whether overfitting of model has occured.
}
