% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_RWNN.R
\name{plot.RWNN}
\alias{plot.RWNN}
\title{Diagnostic-plots of an RWNN-object}
\usage{
\method{plot}{RWNN}(x, ...)
}
\arguments{
\item{x}{An \link{RWNN-object}.}

\item{...}{Additional arguments.}
}
\description{
Diagnostic-plots of an RWNN-object
}
\details{
The additional arguments used by the function are '\code{X_val}' and '\code{y_val}', i.e. the features and targets of the validation-set. These are helpful when analysing whether overfitting of model has occured.
}
