% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R
\name{predict.BRVFL}
\alias{predict.BRVFL}
\title{Predicting targets of an BRVFL object.}
\usage{
\method{predict}{BRVFL}(object, ...)
}
\arguments{
\item{object}{An BRVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Predicting targets of an BRVFL object.
}
\details{
The additional argument '\code{newdata}' and '\code{type}' can be specified:
\describe{
  \item{\code{newdata}}{Expects a matrix the same number of features (columns) as in the original data.}
  \item{\code{type}}{Takes values \code{"all"}, \code{"sd"}, and \code{"mean"} (default), returning a full matrix of predictions for all bootstrap samples, the standard deviation of each predicted observation across bootstrap samples, and the average value of each prediction across the bootstrap samples, respectively.}
}
}
