% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFRVFL.R, R/RVFL.R
\name{predict.RFRVFL}
\alias{predict.RFRVFL}
\alias{predict.RVFL}
\title{Predicting targets of an RFRVFL object.}
\usage{
\method{predict}{RFRVFL}(object, ...)

\method{predict}{RVFL}(object, ...)
}
\arguments{
\item{object}{An RVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Predicting targets of an RFRVFL object.

Predicting targets of an RVFL object.
}
