% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R, R/RVFL.R
\name{predict.BRVFL}
\alias{predict.BRVFL}
\alias{predict.RVFL}
\title{Predicting targets of an BRVFL object.}
\usage{
\method{predict}{BRVFL}(object, ...)

\method{predict}{RVFL}(object, ...)
}
\arguments{
\item{object}{An RVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Predicting targets of an BRVFL object.

Predicting targets of an RVFL object.
}
