% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_RWNN.R
\name{predict.RWNN}
\alias{predict.RWNN}
\title{Predicting targets of an RWNN-object}
\usage{
\method{predict}{RWNN}(object, ...)
}
\arguments{
\item{object}{An \link{RWNN-object}.}

\item{...}{Additional arguments.}
}
\value{
A vector of predicted values.
}
\description{
Predicting targets of an RWNN-object
}
\details{
The additional arguments used by the function are '\code{newdata}' and '\code{class}'. The argument '\code{newdata}' expects a \link{matrix} or \link{data.frame} with the same features (columns) as in the original data. While the '\code{class}' argument can be set to \code{"classify"}. If \code{class == "classify"} additional arguments '\code{t}' and '\code{b}' can be passed to the \link{classify}-function.
}
