% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_network.R
\name{reduce_network}
\alias{reduce_network}
\alias{reduce_network.RWNN}
\alias{reduce_network.ERWNN}
\title{Reduce the weights of a random weight neural network.}
\usage{
reduce_network(object, method = NULL, p = 0.1, ...)

\method{reduce_network}{RWNN}(object, method = NULL, p = 0.1, ...)

\method{reduce_network}{ERWNN}(object, method = NULL, p = 0.1, ...)
}
\arguments{
\item{object}{An \link{RWNN} or \link{ERWNN}-object.}

\item{method}{A string setting the method used to reduce the network (see details).}

\item{p}{The fraction indicating which proportion of weights should be removed during reduction.}

\item{...}{Additional arguments passed to the reduction method.}
}
\value{
A reduced \link{RWNN} or \link{ERWNN}-object.
}
\description{
Methods for weight and neuron pruning in random weight neural networks.
}
\details{
...
}
\examples{
m <- rwnn(y ~ ., data = example_data, N_hidden = 10, lambda = 200, control = list(lnorm = "l1"))

m |> 
    reduce_network(method = "lamp", p = 0.1) |> 
    (\(x) x$Weights)()

m |> 
    reduce_network(method = "lamp", p = 0.1) |>
    reduce_network(method = "last") |> 
    (\(x) x$Weights)()
}
