% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_network.R
\name{reduce_network}
\alias{reduce_network}
\alias{reduce_network.RWNN}
\alias{reduce_network.ERWNN}
\title{Reduce the weights of a random weight neural network.}
\usage{
reduce_network(object, method, retrain = TRUE, ...)

\method{reduce_network}{RWNN}(object, method, retrain = TRUE, ...)

\method{reduce_network}{ERWNN}(object, method, retrain = TRUE, ...)
}
\arguments{
\item{object}{An \link{RWNN} or \link{ERWNN}-object.}

\item{method}{A string setting the method, or a function, used to reduce the network  (see details).}

\item{retrain}{TRUE/FALSE: Should the output weights be retrained after reduction?}

\item{...}{Additional arguments passed to the reduction method (see details).}
}
\value{
A reduced \link{RWNN} or \link{ERWNN}-object.
}
\description{
Methods for weight and neuron pruning in random weight neural networks.
}
\details{
...
}
\examples{
## RWNN
\dontrun{
m <- rwnn(y ~ ., data = example_data, N_hidden = c(10, 15), lambda = 2, control = list(lnorm = "l2"))

m |> 
    reduce_network(method = "relief", p = 0.2, type = "neuron") |> 
    (\(x) x$Weights)()

m |> 
    reduce_network(method = "relief", p = 0.2, type = "neuron") |> 
    reduce_network(method = "correlationtest", rho = 0.995, alpha = 0.05) |> 
    (\(x) x$Weights)()


m |> 
    reduce_network(method = "relief", p = 0.2, type = "neuron") |> 
    reduce_network(method = "correlationtest", rho = 0.995, alpha = 0.05) |> 
    reduce_network(method = "lamp", p = 0.2) |> 
    (\(x) x$Weights)()

m |> 
    reduce_network(method = "relief", p = 0.4, type = "neuron") |> 
    reduce_network(method = "relief", p = 0.4, type = "weight") |> 
    reduce_network(method = "output") |> 
    (\(x) x$Weights)()
}

## ERWNN (reduction is performed element-wise on each RWNN)
\dontrun{
m <- bag_rwnn(y ~ ., data = example_data, N_hidden = c(10, 15), lambda = 2, B = 100, control = list(lnorm = "l2"))

m |> 
    reduce_network(method = "relief", p = 0.2, type = "neuron") |> 
    reduce_network(method = "relief", p = 0.2, type = "weight") |> 
    reduce_network(method = "output")
}
}
