% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R
\name{residuals.BRVFL}
\alias{residuals.BRVFL}
\title{Residuals of the BRVFL object.}
\usage{
\method{residuals}{BRVFL}(object, ...)
}
\arguments{
\item{object}{An BRVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Residuals of the BRVFL object.
}
\details{
No additional arguments are used in this instance.
}
