% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUX_ERVFL.R
\name{residuals.ERVFL}
\alias{residuals.ERVFL}
\title{Residuals of the ERVFL object.}
\usage{
\method{residuals}{ERVFL}(object, ...)
}
\arguments{
\item{object}{An ERVFL-object.}

\item{...}{Additional arguments.}
}
\value{
A vector of raw residuals between the predicted (using \code{type = "mean"}) and observed targets.
}
\description{
Residuals of the ERVFL object.
}
\details{
No additional arguments are used in this instance.
}
