% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFRVFL.R, R/RVFL.R
\name{residuals.RFRVFL}
\alias{residuals.RFRVFL}
\alias{residuals.RVFL}
\title{Residuals of the RFRVFL object.}
\usage{
\method{residuals}{RFRVFL}(object, ...)

\method{residuals}{RVFL}(object, ...)
}
\arguments{
\item{object}{An RVFL-object.}

\item{...}{Additional arguments.}
}
\description{
Residuals of the RFRVFL object.

Residuals of the RVFL object.
}
