% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFRVFL-package.R, R/RFRVFL.R
\docType{package}
\name{RFRVFL}
\alias{RFRVFL}
\alias{RFRVFL.default}
\title{RFRVFL: Random Forest Random Vector Functional Link Neural Network Models}
\usage{
RFRVFL(X, y, N_hidden, B = NULL, ...)

\method{RFRVFL}{default}(X, y, N_hidden, B, ...)
}
\arguments{
\item{X}{A matrix of observed features used to estimate the parameters of the output layer.}

\item{y}{A vector of observed targets used to estimate the parameters of the output layer.}

\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{B}{The number of bootstrap samples.}

\item{...}{Additional arguments passed to the \link{control_RVFL} function.}
}
\value{
An RFRVFL-object containing the random and fitted weights of all bootstrapped RVFL-model.
}
\description{
This package allows for simple creation, estimation, and prediction of random vector functional link neural networks. Furthermore, it incorporates a random forest style boosting into the RVFL framework, yielding more robust estimation and prediction.

Set-up and estimate weights the ensemble random forest random vector functional link neural network.
}
\examples{
N <- 1000
p <- 5

X <- matrix(rnorm(N), ncol = p) 
beta <- matrix(runif(p), ncol = 1) 
y <- X \%*\% beta + rnorm(p)

N_hidden <- c(10, 2, 4, 2)
B <- 1000
RVFL(X = X, y = y, N_hidden = N_hidden, B = B, combine_input = FALSE)
}
\author{
Søren B. Vilsen <svilsen@math.aau.dk>
}
