% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVFL.R
\name{RVFL}
\alias{RVFL}
\alias{RVFL.default}
\title{Random vector functional link}
\usage{
RVFL(X, y, N_hidden, ...)

\method{RVFL}{default}(X, y, N_hidden, ...)
}
\arguments{
\item{X}{A matrix of observed features used to estimate the parameters of the output layer.}

\item{y}{A vector of observed targets used to estimate the parameters of the output layer.}

\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{...}{Additional arguments passed to the \link{control_RVFL} function.}
}
\value{
An RVFL-object containing the random and fitted weights of the RVFL-model.
}
\description{
Set-up and estimate weights of a random vector functional link neural network.
}
\examples{
N <- 1000
p <- 5

X <- matrix(rnorm(N), ncol = p) 
beta <- matrix(runif(p), ncol = 1) 
y <- X \%*\% beta + rnorm(p)

N_hidden <- c(10, 2, 4, 2)
RVFL(X = X, y = y, N_hidden = N_hidden, combine_input = FALSE)
}
