% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRVFL.R
\name{sampleRVFL}
\alias{sampleRVFL}
\alias{sampleRVFL.default}
\title{Sampling random vector functional links}
\usage{
sampleRVFL(X, y, N_hidden, control_rvfl = list(), control_sample = list())

\method{sampleRVFL}{default}(X, y, N_hidden, control_rvfl = list(), control_sample = list())
}
\arguments{
\item{X}{A matrix of observed features used to estimate the parameters of the output layer.}

\item{y}{A vector of observed targets used to estimate the parameters of the output layer.}

\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{control_rvfl}{A list of additional arguments passed to the \link{control_RVFL} function.}

\item{control_sample}{A list of additional arguments passed to the \link{control_sampleRVFL} function.}
}
\value{
The return object will depend on the choice of \code{method} passed through the \code{control_sample} argument. 
If \code{method} is set to \code{"map"} an \link{RVFL-object} is returned, while an \link{ERVFL-object} is returned when \code{method} is set to \code{"resample"}.
}
\description{
Uses sampling to pre-train sampling distribution of the hidden layers in random vector functional link neural network models.
}
\examples{
N <- 2000
p <- 5

s <- seq(0, pi, length.out = N)
X <- matrix(NA, ncol = p, nrow = N)
X[, 1] <- sin(s)
X[, 2] <- cos(s)
X[, 3] <- s
X[, 4] <- s^2
X[, 5] <- s^3

beta <- matrix(rnorm(p), ncol = 1) 
y <- X \%*\% beta + rnorm(N, 0, 1)

N_hidden <- 1000

## Returning an RVFL object using just the MAP estimate of the weights
\dontrun{
sampleRVFL(X = X, y = y, N_hidden = N_hidden, control_sample = list(method = "map"))
}

## Returning an ERVFL object resampling weights from the create posterior sample
\dontrun{
sampleRVFL(X = X, y = y, N_hidden = N_hidden, control_sample = list(method = "resample"))
}

}
