% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R
\name{sampleRVFL}
\alias{sampleRVFL}
\alias{sampleRVFL.default}
\title{Sampling random vector functional links}
\usage{
sampleRVFL(
  X,
  y,
  N_hidden,
  N_simulations = 4000,
  N_burnin = 1000,
  N_resample = 100,
  trace = NULL,
  ...
)

\method{sampleRVFL}{default}(
  X,
  y,
  N_hidden,
  N_simulations = 4000,
  N_burnin = 1000,
  N_resample = 100,
  trace = NULL,
  ...
)
}
\arguments{
\item{X}{A matrix of observed features used to estimate the parameters of the output layer.}

\item{y}{A vector of observed targets used to estimate the parameters of the output layer.}

\item{N_hidden}{A vector of integers designating the number of neurons in each of the hidden layers (the length of the list is taken as the number of hidden layers).}

\item{N_simulations}{The number of Metropolis-Hastings samples drawn from the posterior.}

\item{N_burnin}{The number of samples used as burn-in during Metropolis-Hastings sampling.}

\item{N_resample}{The number of samples drawn during re-sampling after Metropolis-Hastings.}

\item{trace}{Show trace every \code{trace} iterations.}

\item{...}{Additional arguments passed to the \link{control_RVFL} function.}
}
\value{
An ERVFL-object containing the following:
\describe{
    \item{\code{data}}{The original data used to estimate the weights.}
    \item{\code{RVFLmodels}}{A list of \link{RVFL}-objects.}
    \item{\code{weights}}{A vector of ensemble weights.}
    \item{\code{method}}{A string indicating the method ('boosting' in this case)}
}
}
\description{
Uses sampling to pre-train sampling distribution of the hidden layers in random vector functional link neural network models.
}
\examples{
N <- 200
p <- 5

X <- matrix(rnorm(N * p), ncol = p) 
beta <- matrix(runif(p), ncol = 1) 
y <- exp(X \%*\% beta) + rnorm(p)

N_hidden <- 100
sampleRVFL(X = X, y = y, N_hidden = N_hidden)
}
