% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R
\name{set_weights}
\alias{set_weights}
\alias{set_weights.BRVFL}
\title{Set ensemble weights for an BRVFL-object.}
\usage{
set_weights(object, weights = NULL)

\method{set_weights}{BRVFL}(object, weights = NULL)
}
\arguments{
\item{object}{An BRVFL-object.}

\item{weights}{A vector of ensemble weights.}
}
\value{
An BRVFL-object.
}
\description{
Manually set ensemble weights for an BRVFL-object.
}
