% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\name{set_weights}
\alias{set_weights}
\alias{set_weights.ERVFL}
\title{Set ensemble weights for an ERVFL-object.}
\usage{
set_weights(object, weights = NULL)

\method{set_weights}{ERVFL}(object, weights = NULL)
}
\arguments{
\item{object}{An ERVFL-object.}

\item{weights}{A vector of ensemble weights.}
}
\value{
An ERVFL-object.
}
\description{
Manually set ensemble weights for an ERVFL-object.
}
\examples{
N <- 200
p <- 5

X <- matrix(rnorm(N * p), ncol = p) 
beta <- matrix(runif(p), ncol = 1) 
y <- X \%*\% beta + rnorm(p)

N_hidden <- 10
B <- 100
mm <- bagRVFL(X = X, y = y, N_hidden = N_hidden, B = B, 
              lambda = 0.2, combine_input = FALSE)

w <- runif(B)
w <- w / sum(w)
set_weights(mm, weights = w)
}
