% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{set_weights}
\alias{set_weights}
\alias{set_weights.ERWNN}
\title{Set ensemble weights of an ERWNN-object}
\usage{
set_weights(object, weights)

\method{set_weights}{ERWNN}(object, weights)
}
\arguments{
\item{object}{An \link{ERWNN-object}.}

\item{weights}{A vector of ensemble weights.}
}
\value{
An \link{ERWNN-object}.
}
\description{
Manually set ensemble weights of an \link{ERWNN-object}.
}
\examples{
\dontrun{
N_hidden <- 10
B <- 50

m <- bag_rwnn(y ~ ., data = example_data, n_hidden = n_hidden, 
               lambda = 0.2, B = B)

w <- runif(B)
w <- w / sum(w)
m <- m |> set_weights(weights = w)
}
}
