% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_ERWNN.R
\name{set_weights}
\alias{set_weights}
\alias{set_weights.ERWNN}
\title{Set ensemble weights for an ERWNN-object}
\usage{
set_weights(object, weights = NULL)

\method{set_weights}{ERWNN}(object, weights = NULL)
}
\arguments{
\item{object}{An \link{ERWNN-object}.}

\item{weights}{A vector of ensemble weights.}
}
\value{
An \link{ERWNN-object}.
}
\description{
Manually set ensemble weights for an \link{ERWNN-object}.
}
\examples{
N <- 2000
p <- 5

s <- seq(0, pi, length.out = N)
X <- matrix(NA, ncol = p, nrow = N)
X[, 1] <- sin(s)
X[, 2] <- cos(s)
X[, 3] <- s
X[, 4] <- s^2
X[, 5] <- s^3

beta <- matrix(rnorm(p), ncol = 1) 
y <- X \%*\% beta + rnorm(N, 0, 1)

N_hidden <- 10
B <- 50
mm <- bag_rwnn(y ~ X, N_hidden = N_hidden,
              lambda = 0.2, B = B)

w <- runif(B)
w <- w / sum(w)
set_weights(mm, weights = w)
}
