% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions_ERWNN.R
\name{set_weights}
\alias{set_weights}
\alias{set_weights.ERWNN}
\title{Set ensemble weights for an ERWNN-object}
\usage{
set_weights(object, weights = NULL)

\method{set_weights}{ERWNN}(object, weights = NULL)
}
\arguments{
\item{object}{An \link{ERWNN-object}.}

\item{weights}{A vector of ensemble weights.}
}
\value{
An \link{ERWNN-object}.
}
\description{
Manually set ensemble weights for an \link{ERWNN-object}.
}
\examples{
N_hidden <- 10
B <- 50
mm <- bag_rwnn(y ~ ., data = example_data, 
               N_hidden = N_hidden, lambda = 0.2, B = B)

w <- runif(B)
w <- w / sum(w)
set_weights(mm, weights = w)
}
