% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRVFL.R
\name{set_weights}
\alias{set_weights}
\alias{set_weights.BRVFL}
\title{Set ensemble weights for an BRVFL-object.}
\usage{
set_weights(object, weights = NULL)

\method{set_weights}{BRVFL}(object, weights = NULL)
}
\arguments{
\item{object}{An BRVFL-object.}

\item{weights}{A vector of ensemble weights.}
}
\value{
An BRVFL-object.
}
\description{
Manually set ensemble weights for an BRVFL-object.
}
\examples{
N <- 200
p <- 5

X <- matrix(rnorm(N * p), ncol = p) 
beta <- matrix(runif(p), ncol = 1) 
y <- X \%*\% beta + rnorm(p)

N_hidden <- c(10, 2, 4, 2)
B <- 100
mm <- BRVFL(X = X, y = y, N_hidden = N_hidden, B = B, combine_input = FALSE)

w <- runif(B)
w <- w / sum(w)
set_weights(mm, weights = w)
}
