% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuningHyperparameters.R
\name{tune_hyperparameters}
\alias{tune_hyperparameters}
\title{Hyper-parameter tuning}
\usage{
tune_hyperparameters(
  method,
  X,
  y,
  folds = 10,
  hyperparameters = list(),
  control = list(),
  trace = 0
)
}
\arguments{
\item{method}{}

\item{X}{A matrix of observed features used to train the parameters of the output layer.}

\item{y}{A vector of observed targets used to train the parameters of the output layer.}

\item{folds}{The number of folds used in k-fold cross-validation.}

\item{hyperparameters}{A named list of hyper-parameters.}

\item{control}{A list of additional arguments passed to the \link{control_RVFL} function.}

\item{trace}{A numeric indicating how often a trace of should be shown (default is 0).}
}
\value{
An object either of class \link{RVFL} or \link{ERVFL}.
}
\description{
Simple function for hyper-parameter tuning using k-fold cross-validation.
}
