% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuningHyperparameters.R
\name{tune_hyperparameters}
\alias{tune_hyperparameters}
\title{Random vector functional link}
\usage{
tune_hyperparameters(method, X, y, folds = 10, control = list(), ...)
}
\arguments{
\item{method}{}

\item{X}{A matrix of observed features used to train the parameters of the output layer.}

\item{y}{A vector of observed targets used to train the parameters of the output layer.}

\item{folds}{The number of folds used in k-fold cross-validation.}

\item{control}{A list of additional arguments passed to the \link{control_RVFL} function.}

\item{...}{}
}
\value{
An object either of \link{RVFL} or \link{ERVFL}.
}
\description{
Set-up and estimate weights of a random vector functional link neural network.
}
